Decided to change the emoji for night-time
It didn't really make sense that I was using a night-time emoji while showing the weather with an emoji with a sun on it. That's just wack. So I decided to change it to the "city sunset" emoji but I didn't like the look of it as much as "city sunrise" so I decided on city sunrise. I mean, that emoji can totally be a sunset too, right? The only people who will know are the ones looking at this source code and also if you hover over the emoji when you're on Twitter in a desktop browser, but who the hell browses Twitter on desktop anyways? Whatever I'm commiting this.